move FakeIdentityManager up a level
